requirement.txt
asgiref==3.2.10
Django==3.1.1
Pillow==7.2.0
python-slugify==4.0.1
pytz==2020.1
sqlparse==0.3.1
text-unidecode==1.3
